user_guide: fix method definitions for string and array
The internal StrType is meant to represent strings of Unicode
characters, in whatever encoding scheme the target language supports.
Therefore the intention is for length and substring methods to work with
characters (or Unicode codes) rather than bytes.

The size() method on arrays is meant to return the number of elements in
the array, not the number of bytes required to store the full array.