Update maven dependency to correct version.
I was trying to adopt this tool in a Java project, and some classes in
the code generated by `ksc` were not found. I eventually thought to go
to Maven Central and, sure enough, found that there was a newer
release than the one the documentation told me to use.

You might want to consider using a tool like Maven Badges,
https://github.com/softwaremill/maven-badges so that your
documentation file is rendered with an image that shows the actual
latest release version, and which when clicked takes you to a page
showing how to add it as a dependency using many popular JVM-hosted
build tools. I do this on my own project pages, for example:
https://github.com/Deep-Symmetry/beat-link#installing